+++ include/ui/widgets/Button.h
@@ -29,7 +29,7 @@ public:
 
     ~Button() override = default;
 
-    void handle_event(const std::optional<sf::Event> &event, const sf::Vector2f &mouse_pos) override {
+    void handle_event(const std::optional<sf::Event> &event, const sf::Vector2i &mouse_pos) override {
 
     }
